requirements.txt
streamlit
pandas
pdfplumber
PyMuPDF
pytesseract
easyocr
fuzzywuzzy
python-Levenshtein
Pillow
numpy
openpyxl
python-dateutil
xlsxwriter